+++ DataInterface/IClientRepository.cs
@@ -5,7 +5,7 @@ namespace DataInterface
 {
     public interface IClientRepository : IRepository<Client>
     {
-        IEnumerable<User> GetUsersClients(int userId);
-        bool LawyerHasClient(int lawyerId, int clientId);
+        IEnumerable<User> GetUsersClients(long userId);
+        bool LawyerHasClient(long lawyerId, int clientId);
     }
 }
